requirements.txt
tweepy
spacy
python-dotenv